\name{ICC-package}
\alias{ICC-package}
\docType{package}
\title{Facilitating Estimation of the Intraclass Correlation Coefficient}
\description{Assist in the estimation of the Intraclass Correlation Coefficient (ICC) from variance components of a one-way analysis of variance and also estimate the number of individuals or groups necessary to obtain an ICC estimate with a desired confidence interval width.}
\details{
\tabular{ll}{
Package: \tab ICC\cr
Type: \tab Package\cr
Version: \tab 2.4.0\cr
Date: \tab 2022-05-17\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\seealso{\code{ICCest}, \code{Nest}, \code{ICCbare},
\code{effort}
}


