\name{ICC-package}
\alias{ICC-package}
\docType{package}
\title{Facilitating Estimation of the Intraclass Correlation Coefficient}
\description{Assist in the estimation of the Intraclass Correlation Coefficient (ICC) from variance components of a one-way analysis of variance and also estimate the number of individuals or groups necessary to obtain an ICC estimate with a desired confidence interval width.}
\seealso{\code{ICCest}, \code{Nest}, \code{ICCbare}, \code{effort}
}
\author{
\strong{Maintainer}: Matthew Wolak \email{matthewwolak@gmail.com} (\href{https://orcid.org/0000-0002-7962-0071}{ORCID})
}

