% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\docType{package}
\name{ICC-package}
\alias{ICC-package}
\alias{_PACKAGE}
\title{Facilitating Estimation of the Intraclass Correlation Coefficient}
\description{
Assist in the estimation of the Intraclass Correlation Coefficient (ICC) from
variance components of a one-way analysis of variance and also estimate the
desired confidence interval width.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewwolak/ICC}
  \item Report bugs at \url{https://github.com/matthewwolak/ICC/issues}
}

}
\author{
\strong{Maintainer}: Matthew Wolak \email{matthewwolak@gmail.com}

}
