\name{ICC-package}
\alias{ICC-package}
\docType{package}
\title{Intraclass Correlation Coefficient}
\description{Functions to use when estimating the Intraclass Correlation Coefficient (ICC)}
\details{
\tabular{ll}{
Package: \tab ICC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-31\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
%\keyword{ package }
%\seealso{}


