\name{ICC-package}
\alias{ICC-package}
\docType{package}
\title{Intraclass Correlation Coefficient}
\description{Functions to use when estimating the Intraclass Correlation Coefficient (ICC)}
\details{
\tabular{ll}{
Package: \tab ICC\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-03-10\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\seealso{\code{ICCest}, \code{Nest}, \code{ICCbare},
\code{effort}
}


